Major Update
A lot of things have been updated. First of all, we have added a light mode. Secondly, when we used to click on notes, it used to redirect us on a Google Drive. And when we clicked on previous year questions, it would redirect us to a onedrive link We have changed that and added all the nodes on our website itself so that we are not directed to any other website. And the user stays on our website as long as possible.